company settings UI done
